+++ packages/xchain-thorchain-amm/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.0.1.0-beta (2022-08-15)
+
+- resolved several issues
+- added avax client
+- added ThorchainCache to manage caching thorchain state
+
 # v0.0.1.0-alpha3 (2022-08-08)
 
 ## Remove